get all users added
